Changed samples from doubles to floats throughout. Reworked IGenerator to have an intent of statefulness, adding methods Reset and Clone and allowing the caller to supply the Clip buffer to be filled. Implemented these changes in all the IGenerator classes. Added class Cloner.cs based on a Stack Overflow article to clone Random objects including internal state. Added a NuGet reference to Newtonsoft.Json. Replaced ClipWaveProvider.cs with GeneratorSampleProvider.cs. Updated MainWindow.xaml.cs to stash the IGenerator instead of a Clip, and to call Reset after generating the visualization. Updated cmdPlay_Click to initialize the IWavePlayer using GeneratorSampleProvider. Added new constant MiddleC to Constants.cs. Added new generator RepeaterGenerator.cs leveraging the ability to reset IGenerators, and added new testing/sample code to cmdGenerate_Click in MainWindows.xaml.cs.

+++ EnvelopePoint.cs
@@ -5,15 +5,15 @@ namespace Synthesizer
 	public struct EnvelopePoint
 	{
 		public int Index;
-		public double Value;
+		public float Value;
 
-		public EnvelopePoint(int index, double value)
+		public EnvelopePoint(int index, float value)
 		{
 			Index = index;
 			Value = value;
 		}
 
-		public static EnvelopePoint AtTime(double time, double value = 0.0)
+		public static EnvelopePoint AtTime(double time, float value = 0.0f)
 			=> new EnvelopePoint(
 				(int)Math.Round(time * Global.SampleRate),
 				value);
